+++ Robust.Shared/Serialization/TypeSerializers/Implementations/Box2Serializer.cs
@@ -14,16 +14,13 @@ using Robust.Shared.Utility;
 namespace Robust.Shared.Serialization.TypeSerializers.Implementations
 {
     [TypeSerializer]
-    public sealed class Box2Serializer : ITypeSerializer<Box2, ValueDataNode>, ITypeCopyCreator<Box2>
+    public sealed class Box2Serializer :
+        ITypeSerializer<Box2, ValueDataNode>,
+        ITypeCopyCreator<Box2>,
+        ITypeSerializer<Box2i, ValueDataNode>,
+        ITypeCopyCreator<Box2i>
     {
-        internal static void NextOrThrow(
-            ref ReadOnlySpan<char> source,
-            out ReadOnlySpan<char> splitValue,
-            string errValue)
-        {
-            if (!SpanSplitExtensions.SplitFindNext(ref source, ',', out splitValue))
-                throw new InvalidMappingException($"Could not parse {nameof(Box2)}: '{errValue}'");
-        }
+        #region Box2
 
         public Box2 Read(ISerializationManager serializationManager, ValueDataNode node,
             IDependencyCollection dependencies,
@@ -48,9 +45,21 @@ namespace Robust.Shared.Serialization.TypeSerializers.Implementations
             return new Box2(l, b, r, t);
         }
 
-        public ValidationNode Validate(ISerializationManager serializationManager, ValueDataNode node,
-            IDependencyCollection dependencies,
+        public DataNode Write(ISerializationManager serializationManager, Box2 value,
+            IDependencyCollection dependencies, bool alwaysWrite = false,
             ISerializationContext? context = null)
+        {
+            var nodeValue =
+                $"{value.Left.ToString(CultureInfo.InvariantCulture)}," +
+                $"{value.Bottom.ToString(CultureInfo.InvariantCulture)}," +
+                $"{value.Right.ToString(CultureInfo.InvariantCulture)}," +
+                $"{value.Top.ToString(CultureInfo.InvariantCulture)}";
+
+            return new ValueDataNode(nodeValue);
+        }
+
+        ValidationNode ITypeValidator<Box2, ValueDataNode>.Validate(ISerializationManager serializationManager, ValueDataNode node,
+            IDependencyCollection dependencies, ISerializationContext? context)
         {
             var args = node.Value.Split(',');
 
@@ -67,9 +76,40 @@ namespace Robust.Shared.Serialization.TypeSerializers.Implementations
                 : new ErrorNode(node, "Failed parsing values of Box2.");
         }
 
-        public DataNode Write(ISerializationManager serializationManager, Box2 value,
-            IDependencyCollection dependencies, bool alwaysWrite = false,
+        [MustUseReturnValue]
+        public Box2 CreateCopy(ISerializationManager serializationManager, Box2 source,
+            SerializationHookContext hookCtx,
             ISerializationContext? context = null)
+        {
+            return new Box2(source.Left, source.Bottom, source.Right, source.Top);
+        }
+
+        #endregion
+
+        #region Box2i
+
+        public Box2i Read(ISerializationManager serializationManager, ValueDataNode node, IDependencyCollection dependencies,
+            SerializationHookContext hookCtx, ISerializationContext? context = null, ISerializationManager.InstantiationDelegate<Box2i>? instanceProvider = null)
+        {
+            var nodeValue = node.Value.AsSpan();
+            NextOrThrow(ref nodeValue, out var current, node.Value);
+
+            var l = Parse.Int32(current);
+            NextOrThrow(ref nodeValue, out current, node.Value);
+
+            var b = Parse.Int32(current);
+            NextOrThrow(ref nodeValue, out current, node.Value);
+
+            var r = Parse.Int32(current);
+            NextOrThrow(ref nodeValue, out current, node.Value);
+
+            var t = Parse.Int32(current);
+
+            return new Box2i(l, b, r, t);
+        }
+
+        public DataNode Write(ISerializationManager serializationManager, Box2i value, IDependencyCollection dependencies,
+            bool alwaysWrite = false, ISerializationContext? context = null)
         {
             var nodeValue =
                 $"{value.Left.ToString(CultureInfo.InvariantCulture)}," +
@@ -80,12 +120,39 @@ namespace Robust.Shared.Serialization.TypeSerializers.Implementations
             return new ValueDataNode(nodeValue);
         }
 
-        [MustUseReturnValue]
-        public Box2 CreateCopy(ISerializationManager serializationManager, Box2 source,
-            SerializationHookContext hookCtx,
+        ValidationNode ITypeValidator<Box2i, ValueDataNode>.Validate(ISerializationManager serializationManager, ValueDataNode node,
+            IDependencyCollection dependencies, ISerializationContext? context)
+        {
+            var args = node.Value.Split(',');
+
+            if (args.Length != 4)
+            {
+                return new ErrorNode(node, "Invalid amount of args for Box2i.");
+            }
+
+            return Parse.TryInt32(args[0], out _) &&
+                   Parse.TryInt32(args[1], out _) &&
+                   Parse.TryInt32(args[2], out _) &&
+                   Parse.TryInt32(args[3], out _)
+                ? new ValidatedValueNode(node)
+                : new ErrorNode(node, "Failed parsing values of Box2i.");
+        }
+
+        public Box2i CreateCopy(ISerializationManager serializationManager, Box2i source, SerializationHookContext hookCtx,
             ISerializationContext? context = null)
         {
-            return new Box2(source.Left, source.Bottom, source.Right, source.Top);
+            return new Box2i(source.Left, source.Bottom, source.Right, source.Top);
+        }
+
+        #endregion
+
+        private static void NextOrThrow(
+            ref ReadOnlySpan<char> source,
+            out ReadOnlySpan<char> splitValue,
+            string errValue)
+        {
+            if (!SpanSplitExtensions.SplitFindNext(ref source, ',', out splitValue))
+                throw new InvalidMappingException($"Could not parse {nameof(Box2)}: '{errValue}'");
         }
     }
 }
